ENH: Unify src use
