+++ mne/source_estimate.py
@@ -12,13 +12,14 @@ import warnings
 
 import numpy as np
 from scipy import linalg, sparse
-from scipy.sparse import csr_matrix, coo_matrix
+from scipy.sparse import coo_matrix
 
 from .filter import resample
 from .evoked import _get_peak
 from .parallel import parallel_func
 from .surface import (read_surface, _get_ico_surface, read_morph_map,
-                      _compute_nearest)
+                      _compute_nearest, mesh_edges)
+from .source_space import _ensure_src
 from .utils import (get_subjects_dir, _check_subject, logger, verbose,
                     _time_mask)
 from .viz import plot_source_estimates
@@ -1798,6 +1799,7 @@ class MixedSourceEstimate(_BaseSourceEstimate):
         """
 
         # extract surface source spaces
+        src = _ensure_src(src)
         surf = [s for s in src if s['type'] == 'surf']
         if len(surf) != 2:
             raise ValueError('Source space must contain exactly two surfaces.')
@@ -1821,64 +1823,6 @@ class MixedSourceEstimate(_BaseSourceEstimate):
 
 ###############################################################################
 # Morphing
-
-
-def mesh_edges(tris):
-    """Returns sparse matrix with edges as an adjacency matrix
-
-    Parameters
-    ----------
-    tris : array of shape [n_triangles x 3]
-        The triangles.
-
-    Returns
-    -------
-    edges : sparse matrix
-        The adjacency matrix.
-    """
-    if np.max(tris) > len(np.unique(tris)):
-        raise ValueError('Cannot compute connectivity on a selection of '
-                         'triangles.')
-
-    npoints = np.max(tris) + 1
-    ones_ntris = np.ones(3 * len(tris))
-
-    a, b, c = tris.T
-    x = np.concatenate((a, b, c))
-    y = np.concatenate((b, c, a))
-    edges = coo_matrix((ones_ntris, (x, y)), shape=(npoints, npoints))
-    edges = edges.tocsr()
-    edges = edges + edges.T
-    return edges
-
-
-def mesh_dist(tris, vert):
-    """Compute adjacency matrix weighted by distances
-
-    It generates an adjacency matrix where the entries are the distances
-    between neighboring vertices.
-
-    Parameters
-    ----------
-    tris : array (n_tris x 3)
-        Mesh triangulation
-    vert : array (n_vert x 3)
-        Vertex locations
-
-    Returns
-    -------
-    dist_matrix : scipy.sparse.csr_matrix
-        Sparse matrix with distances between adjacent vertices
-    """
-    edges = mesh_edges(tris).tocoo()
-
-    # Euclidean distances between neighboring vertices
-    dist = np.sqrt(np.sum((vert[edges.row, :] - vert[edges.col, :]) ** 2,
-                          axis=1))
-
-    dist_matrix = csr_matrix((dist, (edges.row, edges.col)), shape=edges.shape)
-
-    return dist_matrix
 
 
 @verbose
